+++ packages/salesforcedx-vscode-apex/src/commands/forceApexTestRunCodeAction.ts
@@ -12,6 +12,7 @@ import {
 } from '@salesforce/salesforcedx-utils-vscode/out/src/cli';
 import * as vscode from 'vscode';
 import { nls } from '../messages';
+import { forceApexTestRunCacheService, isEmpty } from '../testRunCache';
 
 const sfdxCoreExports = vscode.extensions.getExtension(
   'salesforce.salesforcedx-vscode-core'
@@ -22,74 +23,6 @@ const SfdxCommandlet = sfdxCoreExports.SfdxCommandlet;
 const SfdxWorkspaceChecker = sfdxCoreExports.SfdxWorkspaceChecker;
 const SfdxCommandletExecutor = sfdxCoreExports.SfdxCommandletExecutor;
 const notificationService = sfdxCoreExports.notificationService;
-
-function isEmpty(value: string): boolean {
-  return !value || value.length === 0;
-}
-
-function isNotEmpty(value: string): boolean {
-  return !isEmpty(value);
-}
-
-// cache last test class and test method values to
-// enable re-running w/o command context via built-in LRU
-class ForceApexTestRunCacheService {
-  private lastClassTestParam: string;
-  private lastMethodTestParam: string;
-  private static instance: ForceApexTestRunCacheService;
-
-  public static getInstance() {
-    if (!ForceApexTestRunCacheService.instance) {
-      ForceApexTestRunCacheService.instance = new ForceApexTestRunCacheService();
-    }
-    return ForceApexTestRunCacheService.instance;
-  }
-
-  public constructor() {
-    this.lastClassTestParam = '';
-    this.lastMethodTestParam = '';
-  }
-
-  public getLastClassTestParam(): string {
-    return this.lastClassTestParam;
-  }
-
-  public getLastMethodTestParam(): string {
-    return this.lastMethodTestParam;
-  }
-
-  public hasCachedClassTestParam() {
-    return isNotEmpty(this.lastClassTestParam);
-  }
-
-  public hasCachedMethodTestParam() {
-    return isNotEmpty(this.lastMethodTestParam);
-  }
-
-  public async setCachedClassTestParam(test: string) {
-    // enable then run 'last executed' command so command
-    // added to 'recently used'
-    await vscode.commands.executeCommand(
-      'setContext',
-      'sfdx:has_cached_test_class',
-      true
-    );
-    this.lastClassTestParam = test;
-  }
-
-  public async setCachedMethodTestParam(test: string) {
-    // enable then run 'last executed' command so command
-    // added to 'recently used'
-    await vscode.commands.executeCommand(
-      'setContext',
-      'sfdx:has_cached_test_method',
-      true
-    );
-    this.lastMethodTestParam = test;
-  }
-}
-
-export const forceApexTestRunCacheService = ForceApexTestRunCacheService.getInstance();
 
 // build force:apex:test:run w/ given test class or test method
 export class ForceApexTestRunCodeActionExecutor extends SfdxCommandletExecutor<{}> {
