+++ packages/salesforcedx-vscode-apex/src/testRunCache/index.ts
@@ -4,11 +4,7 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-
 export {
-  forceApexTestClassRunCodeAction,
-  forceApexTestClassRunCodeActionDelegate,
-  forceApexTestMethodRunCodeAction,
-  forceApexTestMethodRunCodeActionDelegate,
-  ForceApexTestRunCodeActionExecutor
-} from './forceApexTestRunCodeAction';
+  forceApexTestRunCacheService,
+  isEmpty
+} from './forceApexTestRunCacheService';
